requirements.txt
beautifulsoup4==4.11.1
lxml
pyyaml
pyfiglet==0.8.post1
requests==2.28.1
termcolor==2.1.1
websocket_client==1.5.1
openai==0.27.8
pydantic==2.1.1
dulwich==0.21.5
litellm==0.1.365